Refactor: TopNav style 분리 및 폴더구조 변경
