+++ src/common/TopNav/components/Account.jsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
-import { StyledMenuButton } from 'common/components/navigator/styled';
-import styled from 'styled-components';
+
+import { StyledUserName, StyledProfileImage, StyledMenuButton } from '../styled';
 
 const Account = ({ user }) => {
   return (
@@ -34,24 +34,5 @@ const LoginButton = () => {
     </Link>
   );
 };
-
-const StyledProfileImage = styled.img`
-  border-radius: 50%;
-  margin-left: 0.5rem;
-`;
-
-const StyledUserName = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding-left: 1rem;
-  border-radius: 3.5rem;
-  background-color: ${({ theme }) => theme.backgroundTheme4};
-  transition: all 0.3s;
-
-  &:hover {
-    background-color: ${({ theme }) => theme.primary};
-  }
-`;
 
 export default Account;
