+++ src/common/TopNav/index.js
@@ -1,0 +1,3 @@
+import TopNav from "./components/TopNav";
+
+export default TopNav;
