+++ src/common/TopNav/styled/Logo.styled.js
@@ -1,0 +1,7 @@
+import { Link } from 'react-router-dom';
+import styled from 'styled-components';
+
+export const StyledTopNavLogo = styled(Link)`
+display: flex;
+align-items: center;
+`;
