+++ src/common/TopNav/styled/index.js
@@ -1,0 +1,4 @@
+export { StyledProfileImage, StyledUserName } from './Account.styled';
+export { StyledTopNavLogo } from './Logo.styled';
+export { StyledTopNav } from './TopNav.styled';
+export { StyledMenuButton } from './MenuItems.styled'
